Add a swap-out graphite-data check
Checking for actual paging is a better check than monitoring the
amount of swap used -- primarily in that it will alert when paging
occurs.

When the paging stops the alert will clear, so it's important to
check this over a reasonable period of time - in this instance the
`--from -30mins` should ensure that we are in an error state for at
least 30 mins.

I've set the defaults to be '> 0' for warning, '> 10' for critical,
mostly because we will want to be aware of any paging activity. We
may want to adjust this to be less sensitive later.