Attempt to use num_cpus grain to set load thresholds
Working on the principle that the load is the number of processes
on the run queue, so should in many cases always be less than the
number of CPUs in a system.

This is not always true - since other factors can affect load - but
it is a decent default.

NB: short term load thresholds should be *much higher* than long-term
    this was an oversight when initially creating these thresholds, now
    corrected.