Fix up sensu permissions to prevent sensu writes
Having configuration and check definitions owned by sensu
allows the sensu processes to write to them. This is dangerous,
as any security issue in sensu daemons will allow attackers to
rewrite the files.

Better instead to have them owned by root, but with read/exec
permission granted to the sensu group.

In the process, tidy up the permissions so that there is no
world-read, so that we can hide knowledge of our checks from
other processes (aka intruders!)