Add ability to specify client params
Via a 'sensu.client_params' pillar hash, allow specification
of per-client parameters on the Sensu node.

This is typically used to override thresholds, when the following
syntax is used:

   -w :::params.widget.warning|20::: -c :::params.widget.critical|50:::

However, it can be used for `:::replacement:::` in many other places.
Powerful stuff.

NB: This is kept under 'params' for now so we don't accidentially conflict
    with any of the 'standard' top-level client parameters such as `metric_prefix`.