Enable 'metric_name' parameter in hipchat handler
Many of our alerts are based around the value of a graphite
graph. Now we have custom_metric.js in Grafana, we can provide
a link in the Hipchat notifier to the source graph of the event.

This has required the creation of a local update of the community
hipchat.rb handler.