Default handlers
This change removes pagerduty from the default handlers so that you have
to explicitly add it to the handlers list on a check by check basis.

We're making this change to keep our phones quiet in general and only add
alerts to pagerduty on checks that we really care about.