Quote the command in the YAML to make complex command YAML safe again.
There's a weired interaction where the check comes from YAML (via the
pillar), and has to go back out as YAML safe in the SLS file for salt to
parse. This is by no means perfect, but works for every case we have
right now.

The following pillar triggered this bug (and caused a 'YAML render
error' from salt):

    sensu:
      check_definitions:
        app_500_error:
          type: basic
          command: "/etc/sensu/plugins/check-elastic.rb -r 1h -w 5 -c 10 -q '@fields.status: [500 TO 599] AND tags: pvb' -t app_500_errors -s 'Number of 500 alerts over last hour'",
          description: 'App 500 errors over the last 1 hour'
          interval: 300
          subscribers:
            - monitoring.server