Improve docs on what parameters to check-elastic.rb script do.
We have two different forms of checks in here - it might be best to
split those out into two different checks on hindsight as the way we
raise alerts are quite different.

Or failing that we should just improve the validation so that you don't
specify some that are ignored for the check type running.